requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.33.0
numpy==1.24.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.1
rpds-py==0.20.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
zipp==3.20.2
